// apps/mobile/lib/api/model/internalHandlerHttpMessageResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Wakeup API
 * Friend-graph chat backend. See docs/WAKEUP.md for the full spec.
 * OpenAPI spec version: 1.0
 */

export interface InternalHandlerHttpMessageResponse {
  body?: string;
  conversation_id?: string;
  created_at?: string;
  deleted_at?: string;
  edited_at?: string;
  id?: string;
  is_deleted?: boolean;
  reply_to_message_id?: string;
  sender_id?: string;
}